@@ -10,68 +10,77 @@
     "start": "nest start",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/main",
+    "prebuild": "rimraf dist",
+    "build": "nest build",
     "prisma:db-pull": "prisma db pull",
     "prisma:generate": "prisma generate",
     "prisma:migrate": "prisma migrate dev --name init",
-    "prebuild": "rimraf dist",
-    "build": "nest build",
     "format": "prettier --write \"{src,test}/**/*.ts\"",
-    "lint": "eslint --fix \"{src,apps,libs,test}/**/*.ts\" ",
+    "lint": "eslint --fix \"{src,apps,libs,test}/**/*.ts\"",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "jest --config ./test/jest-e2e.json"
-  },
-  "dependencies": {
-    "@nestjs/common": "^8.4.4",
-    "@nestjs/config": "^2.0.0",
-    "@nestjs/core": "^8.4.4",
-    "@nestjs/mapped-types": "1.0.1",
-    "@nestjs/platform-express": "^8.4.4",
-    "@nestjs/swagger": "^5.2.1",
-    "@prisma/client": "^3.13.0",
-    "ali-oss": "^6.17.1",
-    "class-transformer": "^0.5.1",
-    "class-validator": "^0.13.2",
-    "dayjs": "^1.11.1",
-    "ejs": "^3.1.7",
-    "ffmpeg-stream": "^0.7.0",
-    "reflect-metadata": "^0.1.13",
-    "request": "^2.88.2",
-    "rimraf": "^3.0.2",
-    "rxjs": "^7.5.5",
-    "swagger-ui-express": "^4.3.0",
-    "ts-morph": "^14.0.0"
-  },
-  "devDependencies": {
-    "@nestjs/cli": "^8.2.5",
-    "@nestjs/schematics": "^8.0.10",
-    "@nestjs/testing": "^8.4.4",
-    "@types/ali-oss": "^6.16.3",
-    "@types/express": "^4.17.13",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^17.0.31",
-    "@types/passport-local": "^1.0.34",
-    "@types/request": "^2.48.8",
-    "@types/supertest": "^2.0.12",
-    "@typescript-eslint/eslint-plugin": "^5.21.0",
-    "@typescript-eslint/parser": "^5.21.0",
-    "eslint": "^8.14.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "jest": "28.0.3",
-    "prettier": "^2.6.2",
-    "prisma": "^3.13.0",
-    "supertest": "^6.2.3",
-    "ts-jest": "^27.1.4",
-    "ts-loader": "^9.3.0",
-    "ts-node": "^10.7.0",
-    "tsconfig-paths": "^3.14.1",
-    "typescript": "^4.6.4"
+    "test:e2e": "jest --config ./test/jest-e2e.json",
+    "compodoc": "npx @compodoc/compodoc -p tsconfig.json -s"
   },
   "prisma": {
     "schema": "prisma/schema.prisma"
+  },
+  "dependencies": {
+    "@nestjs/common": "^9.1.2",
+    "@nestjs/config": "^2.2.0",
+    "@nestjs/core": "^9.1.2",
+    "@nestjs/platform-express": "^9.1.2",
+    "@prisma/client": "^4.4.0",
+    "ali-oss": "^6.17.1",
+    "axios": "^0.27.2",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.13.2",
+    "crypto-js": "^4.1.1",
+    "dayjs": "^1.11.5",
+    "ejs": "^3.1.8",
+    "ffmpeg-stream": "^0.7.0",
+    "nanoid": "3.3.4",
+    "node-fetch": "^3.2.10",
+    "progress-stream": "^2.0.0",
+    "puppeteer": "^18.0.5",
+    "reflect-metadata": "^0.1.13",
+    "rimraf": "^3.0.2",
+    "rxjs": "^7.5.7",
+    "swagger-ui-express": "^4.5.0"
+  },
+  "devDependencies": {
+    "@compodoc/compodoc": "^1.1.19",
+    "@nestjs/cli": "^9.1.4",
+    "@nestjs/schematics": "^9.0.3",
+    "@nestjs/swagger": "^6.1.2",
+    "@nestjs/testing": "^9.1.2",
+    "@types/ali-oss": "^6.16.4",
+    "@types/crypto-js": "^4.1.1",
+    "@types/express": "^4.17.14",
+    "@types/jest": "29.1.0",
+    "@types/multer": "^1.4.7",
+    "@types/node": "^18.7.23",
+    "@types/progress-stream": "^2.0.2",
+    "@types/supertest": "^2.0.12",
+    "@typescript-eslint/eslint-plugin": "^5.38.1",
+    "@typescript-eslint/parser": "^5.38.1",
+    "eslint": "^8.24.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "express": "^4.18.1",
+    "jest": "^29.1.1",
+    "prettier": "^2.7.1",
+    "prisma": "^4.4.0",
+    "source-map-support": "^0.5.21",
+    "supertest": "^6.2.4",
+    "ts-jest": "^29.0.3",
+    "ts-loader": "^9.4.1",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.1.0",
+    "typescript": "^4.8.4",
+    "webpack": "^5.74.0"
   },
   "jest": {
     "moduleFileExtensions": [
