@@ -1,18 +1,17 @@
-import { CallHandler, ExecutionContext, Injectable, NestInterceptor } from '@nestjs/common'
-import { Observable } from 'rxjs'
-import { map } from 'rxjs/operators'
-import { ResponseType } from './response.type'
+import { CallHandler, ExecutionContext, Injectable, NestInterceptor } from '@nestjs/common';
+import { map } from 'rxjs/operators';
+import { ResType } from '../typings/response.type';
 
 @Injectable()
 export class TranformInterceptor implements NestInterceptor {
-  intercept(context: ExecutionContext, next: CallHandler): Observable<ResponseType<any>> {
+  intercept(context: ExecutionContext, next: CallHandler<ResType>) {
     return next.handle().pipe(
       map(res => ({
-        statusCode: 200,
-        data: res.data || null,
-        message: res.message,
-        error: false,
+        statusCode: res.statusCode || 200,
+        data: res.data ?? (res || null),
+        message: res.message || '',
+        error: res.error || false,
       }))
-    )
+    );
   }
 }
