@@ -1,24 +1,22 @@
-import { Test, TestingModule } from '@nestjs/testing'
-import { AppController } from './app.controller'
-import { AppService } from './app.service'
+import { Test, TestingModule } from '@nestjs/testing';
+import { AppController } from './app.controller';
+import { AppService } from './app.service';
 
 describe('AppController', () => {
-  let appController: AppController
+  let appController: AppController;
 
   beforeEach(async () => {
     const app: TestingModule = await Test.createTestingModule({
       controllers: [AppController],
       providers: [AppService],
-    }).compile()
+    }).compile();
 
-    appController = app.get<AppController>(AppController)
-  })
+    appController = app.get<AppController>(AppController);
+  });
 
   describe('root', () => {
-    it('should return "Hello World!"', () => {
-      expect(appController.root()).toStrictEqual({
-        message: '欢迎回来 NestJs!',
-      })
-    })
-  })
-})
+    it('should return "happy-planet-service！"', () => {
+      expect(appController.getHello().message).toBe('happy-planet-service！');
+    });
+  });
+});
