@@ -1,9 +1,9 @@
-import { Body, Controller, Post } from '@nestjs/common'
-import { ApiOperation, ApiOkResponse, ApiTags } from '@nestjs/swagger'
-import { MediaService } from './media.service'
-import { ResType } from '@/common/response.type'
-import { TranscodeAudioDto } from './dto/media.dto'
-import { TranscodeAudioEntity } from './entities/media.entity'
+import { Body, Controller, Post } from '@nestjs/common';
+import { ApiOperation, ApiOkResponse, ApiTags } from '@nestjs/swagger';
+import { MediaService } from './media.service';
+import { ResType } from '@/typings/response.type';
+import { TranscodeAudioDto } from './dto/media.dto';
+import { UploadEntity } from '@/upload/entities/upload.entity';
 
 @ApiTags('media')
 @Controller('media')
@@ -13,21 +13,14 @@ export class MediaController {
   /**
    * aac文件转mp3
    */
-  @Post('audio/transcode')
-  @ApiOperation({ summary: '音频转码' })
+  @Post('/audioTranscode')
+  @ApiOperation({ summary: 'acc音频转码mp3' })
   @ApiOkResponse({
     status: 200,
-    type: TranscodeAudioEntity,
-    description: '转码mp3后的url',
+    type: String,
+    description: '转码后的url',
   })
-  async transcodeAudio(@Body() body: TranscodeAudioDto): Promise<ResType<TranscodeAudioEntity>> {
-    const resultUrl = await this.mediaService.convert(body.originAudioSrc)
-
-    return {
-      data: {
-        resultUrl,
-      },
-      message: '成功',
-    }
+  async transcodeAudio(@Body() body: TranscodeAudioDto): Promise<ResType<UploadEntity>> {
+    return await this.mediaService.convert(body.originAudioSrc);
   }
 }
