@@ -1,9 +1,10 @@
-import { Module } from '@nestjs/common'
-import { MediaController } from './media.controller'
-import { MediaService } from './media.service'
+import { FileManagerService } from '@/file-manager/file-manager.service';
+import { Module } from '@nestjs/common';
+import { MediaController } from './media.controller';
+import { MediaService } from './media.service';
 
 @Module({
   controllers: [MediaController],
-  providers: [MediaService],
+  providers: [MediaService, FileManagerService],
 })
 export class MediaModule {}
