@@ -1,23 +1,21 @@
-import { Test, TestingModule } from '@nestjs/testing'
-import { INestApplication } from '@nestjs/common'
-import request from 'supertest'
-import { AppModule } from './../src/app.module'
+import { Test, TestingModule } from '@nestjs/testing';
+import { INestApplication } from '@nestjs/common';
+import request from 'supertest';
+import { AppModule } from './../src/app.module';
 
 describe('AppController (e2e)', () => {
-  let app: INestApplication
+  let app: INestApplication;
 
   beforeEach(async () => {
     const moduleFixture: TestingModule = await Test.createTestingModule({
       imports: [AppModule],
-    }).compile()
+    }).compile();
 
-    app = moduleFixture.createNestApplication()
-    await app.init()
-  })
+    app = moduleFixture.createNestApplication();
+    await app.init();
+  });
 
   it('/ (GET)', () => {
-    return request(app.getHttpServer()).get('/').expect(200).expect({
-      message: '欢迎回来 NestJs!',
-    })
-  })
-})
+    return request(app.getHttpServer()).get('/').expect(200).expect('Welcome to happy-planet-service！');
+  });
+});
